requirements.txt
# # altair==6.0.0
# # attrs==25.4.0
# # beautifulsoup4==4.14.3
# # blinker==1.9.0
# # cachetools==6.2.6
# # certifi==2026.1.4
# # charset-normalizer==3.4.4
# # click==8.3.1
# # colorama==0.4.6
# # contourpy==1.3.3
# # cycler==0.12.1
# # filelock==3.20.3
# # fonttools==4.61.1
# # fsspec==2026.2.0
# # gdown==5.2.1
# # gitdb==4.0.12
# # GitPython==3.1.46
# # idna==3.11
# # Jinja2==3.1.6
# # jsonschema==4.26.0
# # jsonschema-specifications==2025.9.1
# # kiwisolver==1.4.9
# # MarkupSafe==3.0.3
# # matplotlib==3.10.8
# # mpmath==1.3.0
# # narwhals==2.16.0
# # networkx==3.6.1
# # numpy==1.26.4
# # opencv-contrib-python-headless
# # packaging==26.0
# # pandas==2.3.3
# # pillow==12.1.0
# # polars==1.38.1
# # polars-runtime-32==1.38.1
# # protobuf==6.33.5
# # psutil==7.2.2
# # pyarrow==23.0.0
# # pydeck==0.9.1
# # pyparsing==3.3.2
# # PySocks==1.7.1
# # python-dateutil==2.9.0.post0
# # pytz==2025.2
# # PyYAML==6.0.3
# # referencing==0.37.0
# # requests==2.32.5
# # rpds-py==0.30.0
# # scipy==1.17.0
# # setuptools==82.0.0
# # six==1.17.0
# # smmap==5.0.2
# # soupsieve==2.8.3
# # streamlit==1.54.0
# # sympy==1.14.0
# # tenacity==9.1.4
# # toml==0.10.2
# # torch==2.10.0
# # torchvision==0.25.0
# # tornado==6.5.4
# # tqdm==4.67.3
# # typing_extensions==4.15.0
# # tzdata==2025.3
# # ultralytics==8.4.14
# # ultralytics-thop==2.0.18
# # urllib3==2.6.3
# # watchdog==6.0.0


# ultralytics==8.3.0
# streamlit==1.38.0
# opencv-contrib-python-headless
# numpy==1.26.4
# gdown==5.2.1
# pillow==12.1.0


streamlit>=1.38.0
gdown==5.2.1
ultralytics
opencv-python-headless
# Turunkan versi pillow agar mau jalan dengan Streamlit
pillow>=9.5.0,<11.0.0
# Gunakan numpy 1.x karena numpy 2.x sering bikin YOLO/OpenCV crash
numpy<2.0.0
# Library pendukung lainnya (biarkan versi otomatis)
pandas
requests
